# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8092
    description: Generated server url
paths:
  /api/v1/subscriber:
    get:
      tags:
        - fitness-controller-rest
      summary: Check by email if subscriber exists
      operationId: checkIfExistsByEmail
      parameters:
        - name: mail
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Find out if the subscriber exists
          content:
            application/json:
              schema:
                type: boolean
        '404':
          description: Subscriber not found
          content:
            application/json:
              schema:
                type: boolean
    post:
      tags:
        - fitness-controller-rest
      summary: Create a new Subscriber
      operationId: createSubscriber
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestSuscriptorDTO'
        required: true
      responses:
        '200':
          description: Creates a new subscriber
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestSuscriptorDTO'
  /api/v1/subscribers:
    get:
      tags:
        - fitness-controller-rest
      summary: List all subscribers
      operationId: getListOfAllSubscribers
      responses:
        '200':
          description: Retrieves a list of all subscribers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseSuscriptorDTO'
components:
  schemas:
    RequestSuscriptorDTO:
      type: object
      properties:
        mail:
          type: string
          description: Subscriber mail
          example: mandela@gmail.com
        name:
          type: string
          description: Subscriber name
          example: Nelson Mandela
    ResponseSuscriptorDTO:
      type: object
      properties:
        id:
          type: integer
          description: Subscriber ID
          format: int64
          example: 1
        mail:
          type: string
          description: Subscriber email
          example: mandela@gmail.com
        name:
          type: string
          description: Subscriber name
          example: Nelson Mandela